audit: split app bootstrap from server listen
